+++ testbed/tests/polygon_shapes.cpp
@@ -112,8 +112,8 @@ public:
 
 		{
 			float w = 1.0f;
-			float b = w / (2.0f + b2Sqrt(2.0f));
-			float s = b2Sqrt(2.0f) * b;
+			float b = w / (2.0f + b2Sqrt(float32(2.0f)));
+			float s = b2Sqrt(float32(2.0f)) * b;
 
 			b2Vec2 vertices[8];
 			vertices[0].Set(0.5f * s, 0.0f);
